feat: replaced io-ts with effect (#2388)

+++ packages/@liexp/backend/src/queries/actors/fetchActors.query.ts
@@ -1,10 +1,9 @@
 import { fp, pipe } from "@liexp/core/lib/fp/index.js";
 import { type http } from "@liexp/shared/lib/io/index.js";
-import * as O from "fp-ts/lib/Option.js";
+import * as O from "effect/Option";
 import { type ReaderTaskEither } from "fp-ts/lib/ReaderTaskEither.js";
 import * as R from "fp-ts/lib/Record.js";
 import * as S from "fp-ts/lib/string.js";
-import { type Int } from "io-ts";
 import { type DatabaseContext } from "../../context/db.context.js";
 import { type ENVContext } from "../../context/env.context.js";
 import { ActorEntity } from "../../entities/Actor.entity.js";
@@ -12,11 +11,11 @@ import { type DBError } from "../../providers/orm/database.provider.js";
 import { DBService } from "../../services/db.service.js";
 
 const defaultQuery: http.Actor.GetListActorQuery = {
-  ids: O.none,
-  q: O.none,
-  withDeleted: O.none,
-  _end: O.some(20 as Int),
-  _start: O.some(0 as Int),
+  ids: O.none(),
+  q: O.none(),
+  withDeleted: O.none(),
+  _end: O.some(20),
+  _start: O.some(0),
   _order: O.some("DESC"),
   _sort: O.some("createdAt"),
 };
